# 资源/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: Hope and Sparks API
  version: 1.0.0
  description: 'Apifox 导入版接口规范。详细业务说明、数据库落库规则和枚举映射见根目录

    `Hope_and_Sparks_API接口文档.md`。

    '
servers:
- url: http://localhost:8080
  description: 本地开发环境
- url: /
  description: 相对路径
security:
- bearerAuth: []
...
paths:
...
  /api/v1/exercise-sets:
    get:
      tags:
      - 练习与测试
      summary: 获取练习/测试列表
      parameters:
      - $ref: '#/components/parameters/planIdQuery'
      - $ref: '#/components/parameters/nodeIdQuery'
      - $ref: '#/components/parameters/status'
      - $ref: '#/components/parameters/type'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Query | `planId` | 否 | string | 学习计划 ID，对应 study_plan.id |

        | Query | `nodeId` | 否 | string | 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code |

        | Query | `status` | 否 | string | 状态筛选，具体枚举见接口文档数据字典 |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}:
    get:
      tags:
      - 练习与测试
      summary: 获取练习/测试详情
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/questions/{questionId}:
    get:
      tags:
      - 练习与测试
      summary: 获取单题答题详情
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      - $ref: '#/components/parameters/questionId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Path | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/questions/{questionId}/answer:
    put:
      tags:
      - 练习与测试
      summary: 保存单题答案
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      - $ref: '#/components/parameters/questionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                answer:
                  oneOf:
                  - type: string
                  - type: array
                    items:
                      type: string
                  - type: object
                    additionalProperties: true
                  description: 用户答案，按题型可为字符串、数组或对象
                durationSeconds:
                  type: integer
                  description: 耗时或资源总时长，单位秒
                flagged:
                  type: boolean
                  description: 是否标记题目
                attachments:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                  description: 附件列表，对应 sys_oss_file
              required:
              - answer
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Path | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |

        | Body | `answer` | 是 | mixed | 用户答案，按题型可为字符串、数组或对象 |

        | Body | `durationSeconds` | 否 | integer | 耗时或资源总时长，单位秒 |

        | Body | `flagged` | 否 | boolean | 是否标记题目 |

        | Body | `attachments` | 否 | array<object> | 附件列表，对应 sys_oss_file |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/questions/{questionId}/flag:
    put:
      tags:
      - 练习与测试
      summary: 标记或取消标记题目
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      - $ref: '#/components/parameters/questionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                flagged:
                  type: boolean
                  example: true
                  description: 是否标记题目
              required:
              - flagged
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Path | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |

        | Body | `flagged` | 是 | boolean | 是否标记题目 |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/questions/{questionId}/hint:
    post:
      tags:
      - 练习与测试
      summary: 请求 Coach 提示
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      - $ref: '#/components/parameters/questionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                hintLevel:
                  type: integer
                  example: 1
                  description: 提示等级
                currentAnswer:
                  oneOf:
                  - type: string
                  - type: array
                    items:
                      type: string
                  - type: object
                    additionalProperties: true
                  description: 当前答案草稿
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Path | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |

        | Body | `hintLevel` | 否 | integer | 提示等级 |

        | Body | `currentAnswer` | 否 | mixed | 当前答案草稿 |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/submit:
    post:
      tags:
      - 练习与测试
      summary: 提交练习/测试
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                durationSeconds:
                  type: integer
                  description: 耗时或资源总时长，单位秒
                answers:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                  description: 批量作答内容
                forceSubmit:
                  type: boolean
                  example: false
                  description: 是否强制提交
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Body | `durationSeconds` | 否 | integer | 耗时或资源总时长，单位秒 |

        | Body | `answers` | 否 | array<object> | 批量作答内容 |

        | Body | `forceSubmit` | 否 | boolean | 是否强制提交 |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/{exerciseSetId}/questions/{questionId}/code-run:
    post:
      tags:
      - 练习与测试
      summary: 提交代码题即时评测
      parameters:
      - $ref: '#/components/parameters/exerciseSetId'
      - $ref: '#/components/parameters/questionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                language:
                  type: string
                  example: javascript
                  description: 代码语言
                code:
                  type: string
                  description: 验证码或代码内容，按接口上下文区分
                input:
                  type: string
                  description: 自定义输入
                testcaseIds:
                  type: array
                  items:
                    type: string
                  description: 测试用例 ID 列表
              required:
              - language
              - code
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `exerciseSetId` | 是 | string | 练习/测试集合 ID，对应 practice_set.id |

        | Path | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |

        | Body | `language` | 是 | string | 代码语言 |

        | Body | `code` | 是 | string | 验证码或代码内容，按接口上下文区分 |

        | Body | `input` | 否 | string | 自定义输入 |

        | Body | `testcaseIds` | 否 | array<string> | 测试用例 ID 列表 |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
  /api/v1/exercise-sets/attempts/{attemptId}/report:
    get:
      tags:
      - 练习与测试
      summary: 获取练习/测试报告
      parameters:
      - $ref: '#/components/parameters/attemptId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `attemptId` | 是 | string | 作答记录/评估报告 ID，对应 user_question_record.id 或 evaluation_report.id |


        主要数据表：practice_set、practice_set_question、question_bank、user_question_record、evaluation_report'
...
  /api/v1/notifications:
    get:
      tags:
      - 通知
      summary: 获取通知列表
      parameters:
      - $ref: '#/components/parameters/type'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：通知推送服务、user_settings、im_message_read'
  /api/v1/notifications/{notificationId}/read:
    put:
      tags:
      - 通知
      summary: 标记通知已读
      parameters:
      - $ref: '#/components/parameters/notificationId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `notificationId` | 是 | string | 通知 ID |


        主要数据表：通知推送服务、user_settings、im_message_read'
...
  /api/v1/notifications/{notificationId}:
    delete:
      tags:
      - 通知
      summary: 删除通知
      parameters:
      - $ref: '#/components/parameters/notificationId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `notificationId` | 是 | string | 通知 ID |


        主要数据表：通知推送服务、user_settings、im_message_read'
  /api/v1/notifications/{notificationId}/actions:
    post:
      tags:
      - 通知
      summary: 通知动作回调
      parameters:
      - $ref: '#/components/parameters/notificationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                action:
                  type: string
                  description: 动作类型
                payload:
                  type: object
                  additionalProperties: true
                  description: 动作扩展参数
              required:
              - action
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `notificationId` | 是 | string | 通知 ID |

        | Body | `action` | 是 | string | 动作类型 |

        | Body | `payload` | 否 | object | 动作扩展参数 |


        主要数据表：通知推送服务、user_settings、im_message_read'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 页码，从 1 开始
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: 每页数量
...
    type:
      name: type
      in: query
      schema:
        type: string
      description: 类型筛选，具体枚举见接口文档数据字典
    status:
      name: status
      in: query
      schema:
        type: string
      description: 状态筛选，具体枚举见接口文档数据字典
...
    planIdQuery:
      name: planId
      in: query
      schema:
        type: string
      description: 学习计划 ID，对应 study_plan.id
...
    nodeIdQuery:
      name: nodeId
      in: query
      schema:
        type: string
      description: 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code
...
    exerciseSetId:
      name: exerciseSetId
      in: path
      required: true
      schema:
        type: string
      description: 练习/测试集合 ID，对应 practice_set.id
    questionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
      description: 题目 ID，对应 question_bank.id
    attemptId:
      name: attemptId
      in: path
      required: true
      schema:
        type: string
      description: 作答记录/评估报告 ID，对应 user_question_record.id 或 evaluation_report.id
...
    notificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
      description: 通知 ID
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    PageSuccess:
      description: 分页成功
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/PageResult'
...
  schemas:
    ApiResponse:
      type: object
      required:
      - code
      - message
      - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
          oneOf:
          - type: object
            additionalProperties: true
          - type: array
            items: {}
          - type: string
          - type: number
          - type: boolean
        requestId:
          type: string
          example: req_20260505_xxxx
    PageResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 120
        list:
          type: array
          items:
            type: object
            additionalProperties: true